requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django==3.2.9
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
gunicorn==20.1.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
whitenoise==6.4.0
